Box -> Rect
This makes it clear that we are working on a rectangle.
Certain operations, mostly pointInside, require rectangles to make sense and I’ve spent a lot of time debugging why they don’t work only to discover that it’s not really a rectangle.
Need to overhaul Rect to fix this