[BezierParameter] provide LUT functions
For various purposes, callers may want to generate lookup tables that are indexed by a bezier parameter.

The simplest example might be generating a table of BezierEvaluate coordinates across input t (e.g., flattening the curve into line segments).  More expensive applications include collision data for objects that lie in curve coordinates.  In general we’ve found it’s sometimes advantageous to precompute data that will be re-used frequently.

In this patch is introduced a family of conversion functions between bezier parameters and LUT indexes.  Implementing tables themselves are out-of-scope for this library, but simple arrays could be used.

* BCLUTParameterForIndex calculates a bezier parameter for a table index.  This is broadly similar to BCVertexToBezierParameter which is intended for more graphical applications.
* BCLUTIndexForParameter calculates a table index for a bezier parameter.  Since this is lossy, it offers a choice of rounding modes.

Both functions come in variants that allow you to specify a range of interest if you want to avoid allocating LUT space for some regions.

Table indexes are equally weighted.