Selectively work around FB8897889=1 on affected GPUs.
Basically, the workaround in the prior commit breaks apple silicon.

See FB8897889 (AMD bug) and FB8899605 (apple bug).  For now it appears we need to run
different code on each GPU.

While we're waiting for a real fix I will compile this into simulator (only).
We might want to apply it to macOS targets as well but I'm hoping this gets
fixed before it becomes more of an issue.

mt2-108