Avoid use of metal::pow
In FB8904929 we noticed that metal::pow often returns the wrong value on macOS.  This causes incorrect results for some operations like BCCubicSplit (and left/right variants).  It could theoretically affect BCCubicEvaluate, BCCubicEvaluatePrime and BCNormalization as well but it seems a lot less likely for those functions.

This may be related to use of negative base and odd exponent, which ought to be well-defined for metal::pow per Metal Shading Language Specification Table 6.4, but in practice behaves like UB with different results on different GPUs.  Seems to be quite finicky about whether the operands are statically known or the dataflow of the arguments.

The fix is to use explicit multiplication where possible (e.g. integer exponent) and where we think the base might be negative.

Also add test coverage.  This test coverage is known to trip on some AMD and Intel systems.  However some GPUs are known to pass the test even though they experience the issue, so it’s not total.

Due to https://github.com/actions/virtual-environments/issues/1779, there is no CI coverage for this issue.

See also FB8904929, mt2-109.